Digital input to analog pin

+++ app/src/main/java/com/example/kollins/androidemulator/uCInterfaces/DataMemory.java
@@ -20,5 +20,5 @@ public interface DataMemory {
 
     void setPinHandler(Handler pinHandler);
 
-    double getMemoryUsage();
+    int getMemoryUsage();
 }
